Article appendix
Appendix of the article 'An Open Source Application Software to Determine the
Preconsolidation Pressure of Soils in Incremental Loading Oedometer Testing:
pySigmaP' by Exneyder A. Montoya-Araque, A. J. Aparicio-Ortube, David G.
Zapata-Medina and  Luis G. Arboleda-Monsalve (in press).

This script contains the code used to generate the figures related to the
section 'Software Functionalities and Illustrative Examples'.